always run the ensure-gateway-cert Job
Thinking about it, we need to always run this.

If a user switches from http to https, the PreSync hook will not run. We can just preseed it and have it available in case they don't have cert manager.

This also opens the door for us to provide a cert upload button.