Push chart releases to ECR, extend linting to the chart&scripts
This enables chart pushes to ECR, which should be the last step we need
to publish on the Marketplace.

Once this merges and the chart goes live, I will test if I can deploy Orion out of that.

For future releases, this means we will need to increment the version in
Chart.yaml - this is what will trigger the push.

So the workflow will be:
1) get a release branch ready
2) push a version bump to Chart.yaml

Since ECR is immutable and mistakes are a bit of a pain to undo there, I
also thought it's a good moment to bring in some linting I had laying
around - it should keep the release process smooth for us in the future
and help catch any issues early.